refactor fig_panels -> views
