+++ pyhdx/panel/apps.py
@@ -2,7 +2,7 @@ from pyhdx.panel.template import GoldenElvis, ExtendedGoldenTemplate
 from pyhdx.panel.theme import ExtendedGoldenDarkTheme, ExtendedGoldenDefaultTheme
 from pyhdx.panel.controllers import *
 from pyhdx.panel.main_controllers import ComparisonController, PyHDXController
-from pyhdx.panel.fig_panels import *
+from pyhdx.panel.views import *
 from pyhdx.panel.log import get_default_handler
 import sys
 from pyhdx import VERSION_STRING_SHORT
