+++ pyhdx/panel/views.py
